multitarget tracker source moved into a separate folder. Package license info added.
