+++ include/costmap_converter/costmap_to_dynamic_obstacles/multitarget_tracker/defines.h
@@ -1,4 +1,5 @@
-// nach https://github.com/Smorodov/Multitarget-tracker/tree/master/Tracker, GPLv3
+// Based on https://github.com/Smorodov/Multitarget-tracker/tree/master/Tracker, GPLv3
+// Refer to README.md in this directory.
 
 #pragma once
 #include <opencv2/opencv.hpp>
